# 8311163: Parallel: Improve large object handling during evacuation