Please review this change to ParallelGC young generation collection's handling
of large objArrays, to now use the infrastructure provided by JDK-8253237 and
JDK-8337709.  (That's the same infrastructure used by G1 young/mixed
collections.)

Testing: mach5 tier1-5

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8311163](https://bugs.openjdk.org/browse/JDK-8311163): Parallel: Improve large object handling during evacuation (**Enhancement** - P4)


### Reviewers
 * [Albert Mingkun Yang](https://openjdk.org/census#ayang) (@albertnetymk - **Reviewer**) Review applies to [e1db9bca](https://git.openjdk.org/jdk/pull/20720/files/e1db9bcad41d82ce303ebfba051673d580e19539)
 * [Thomas Schatzl](https://openjdk.org/census#tschatzl) (@tschatzl - **Reviewer**)

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/20720/head:pull/20720` \
`$ git checkout pull/20720`

Update a local copy of the PR: \
`$ git checkout pull/20720` \
`$ git pull https://git.openjdk.org/jdk.git pull/20720/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 20720`

View PR using the GUI difftool: \
`$ git pr show -t 20720`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/20720.diff">https://git.openjdk.org/jdk/pull/20720.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/20720#issuecomment-2311334165)