@albertnetymk 2024-08-27
Using `-Xmx3g -XX:+UseParallelGC -XX:ParallelGCThreads=6 DelayInducer.java`, I can see large regression on my box (nproc=12). Have you seen sth similar?

```
# baseline
[0.002s][info][gc] Using Parallel
[1.252s][info][gc] GC(0) Pause Young (Allocation Failure) 381M->296M(479M) 50.696ms
[1.490s][info][gc] GC(1) Pause Young (Allocation Failure) 422M->422M(674M) 165.177ms
[1.864s][info][gc] GC(2) Pause Young (Allocation Failure) 673M->674M(925M) 212.106ms
[2.243s][info][gc] GC(3) Pause Young (Allocation Failure) 925M->925M(1428M) 229.982ms

# new
[0.005s][info][gc] Using Parallel
[5.504s][info][gc] GC(0) Pause Young (Allocation Failure) 381M->299M(479M) 203.409ms
[6.233s][info][gc] GC(1) Pause Young (Allocation Failure) 424M->425M(676M) 620.718ms
[7.594s][info][gc] GC(2) Pause Young (Allocation Failure) 676M->676M(928M) 1155.301ms
[8.920s][info][gc] GC(3) Pause Young (Allocation Failure) 927M->928M(1430M) 1119.259ms
```