@tschatzl 2024-08-29
I did some runs with 1, 2, 4, 6, 12 and 25 parallel workers and I can't see regressions either; the changes seem to be better always. The only difference from your runs is that with `-Xmx3g` there are in total four young collections every time. The last one is the one showing the 2x+ speedup.