@tschatzl 2024-08-29
The old code seems to only ever push a single continuation task, while the new code pushes multiple (after the first one), and tries to keep them at `ParallelGCThreads`.